web/refactor: split up api into server and client
