+++ web/app/page.tsx
@@ -2,9 +2,9 @@ import { ComingSoon } from './components/ComingSoon'
 import { ExplainingTitle } from './components/ExplainingTitle'
 import Hero from './components/Hero'
 import SigmaList from './components/Sigma/SigmaList'
-import SWRConfigProvider from './components/SWRConfigProvider'
+import SWRConfigProvider from './components/api/SWRConfigProvider'
+import { getFallback } from './components/api/serverApi'
 import { MARKET_CAP_BUCKETS } from './misc/MARKET_CAP_BUCKETS'
-import { getFallback } from './misc/tendenzApi'
 
 export default async function Home() {
 	const fallback = await getFallback(
